fonts/tools/requirements.txt
fonttools==4.55.3
PyYAML==6.0.2
